Use Stream ID for stream identification
Up to now, whenever packets or messages needed to be stored by the
stream they belong to, they were sltored by either the source or the
destination MS or both, which is not only a bit unwieldy but also does
not allow for the distinction of multiple streams between a single pair
of mobile stations when they differ, for example, in their period.

+++ src/MsMac.h
@@ -19,7 +19,7 @@ using namespace std;
 
 class MsMac : public cSimpleModule  {
     private:
-	unordered_map<int,cQueue> streamQueues;
+	unordered_map<unsigned long,cQueue> streamQueues;
         Position msPosition;
         int msId;
         int bsId;
